Move downloading images to HTML filter
